+++ source/renderer/app/themes/daedalus/light-blue.js
@@ -333,6 +333,9 @@ export const LIGHT_BLUE_THEME_CONFIG = {
   nodeUpdate: {
     '--theme-node-update-background-color': '#ebeff2',
     '--theme-node-update-title-color': '#5e6066',
+    '--theme-node-sync-info-message-background-color': '#fafbfc',
+    '--theme-node-sync-info-message-text-color': '#5e6066',
+    '--theme-node-sync-icon-color': '#fff',
   },
   notification: {
     '--theme-notification-message-background-color': 'rgba(36, 61, 98, 0.95)',
@@ -637,7 +640,7 @@ export const LIGHT_BLUE_THEME_CONFIG = {
     '--theme-choice-tabs-bottom-border-color-active': '#5e6066',
   },
   testEnvironment: {
-    '--theme-test-environment-label-background-color': '#ab1700',
+    '--theme-test-environment-label-background-color': '#00ab66',
     '--theme-test-environment-label-text-color': '#fafbfc',
   },
   topBar: {