+++ source/renderer/app/themes/daedalus/yellow.js
@@ -455,6 +455,9 @@ export const YELLOW_THEME_CONFIG = {
   nodeUpdate: {
     '--theme-node-update-background-color': '#f8f3ed',
     '--theme-node-update-title-color': '#2d2d2d',
+    '--theme-node-sync-info-message-background-color': '#f8f3ed',
+    '--theme-node-sync-info-message-text-color': '#2d2d2d',
+    '--theme-node-sync-icon-color': '#2d2d2d',
   },
   notification: {
     '--theme-notification-message-background-color': '#fdcd68f2',
@@ -631,7 +634,7 @@ export const YELLOW_THEME_CONFIG = {
     '--theme-choice-tabs-bottom-border-color-active': '#2d2d2d',
   },
   testEnvironment: {
-    '--theme-test-environment-label-background-color': '#ab1700',
+    '--theme-test-environment-label-background-color': '#00ab66',
     '--theme-test-environment-label-text-color': '#fafbfc',
   },
   topBar: {